requirements.txt
anyio==3.7.0
blis==0.7.9
catalogue==2.0.8
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
confection==0.0.4
contourpy==1.0.7
cycler==0.11.0
cymem==2.0.7
exceptiongroup==1.1.1
fastai==2.7.12
fastapi==0.97.0
fastcore==1.5.29
fastdownload==0.0.7
fastprogress==1.0.3
filelock==3.12.1
fonttools==4.40.0
gunicorn==20.1.0
h11==0.14.0
idna==3.4
importlib-resources==5.12.0
Jinja2==3.1.2
joblib==1.2.0
kiwisolver==1.4.4
langcodes==3.3.0
MarkupSafe==2.1.3
matplotlib==3.7.1
mpmath==1.3.0
murmurhash==1.0.9
networkx==3.1
numpy==1.24.3
packaging==23.1
pandas==2.0.2
pathy==0.10.1
Pillow==9.5.0
preshed==3.0.8
pydantic==1.10.9
pyparsing==3.0.9
python-dateutil==2.8.2
python-multipart==0.0.6
pytz==2023.3
PyYAML==6.0
requests==2.31.0
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
smart-open==6.3.0
sniffio==1.3.0
spacy==3.5.3
spacy-legacy==3.0.12
spacy-loggers==1.0.4
srsly==2.4.6
starlette==0.27.0
sympy==1.12
thinc==8.1.10
threadpoolctl==3.1.0
torch==2.0.1
torchvision==0.15.2
tqdm==4.65.0
typer==0.7.0
typing_extensions==4.6.3
tzdata==2023.3
urllib3==2.0.3
uvicorn==0.22.0
wasabi==1.1.2
zipp==3.15.0
